[vtpty] Move UnixPipe to UnixUtils.h and also add UnixUtils.cpp
